phase 1/Mathematics for Computer Vision/Project 4/requirements.txt
numpy
pillow
matplotlib